HGCM,SharedFolders: Added new variation on the HGCM page list type that does not use a bounce buffer.  bugref:9172 [fix]
- Added VMMDevHGCMParmType_NoBouncePageList.
- Made VMMDevHGCMParmType_Embedded and VMMDevHGCMParmType_ContiguousPageList return updated buffer size to the guest.
- Changed the SHFL_FN_READ and SHFL_FN_WRITE function to accept VMMDevHGCMParmType_NoBouncePageList as buffer parameter type, using the RTFileSgReadAt/WriteAt functions to do the work.
- Modified the offset parameter of SHFL_FN_WRITE to return an update offset to assist in the RTFILE_O_APPEND scenarios where other parties have extended the file without the guest knowing.  The guest now gets the file size returned after each write to an handle opened with RTFILE_O_APPEND.



git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@77246 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/HostServices/SharedFolders/vbsf.cpp
@@ -1040,7 +1040,7 @@ static int vbsfPagesToSgBuf(VBOXHGCMSVCPARMPAGES const *pPages, uint32_t cbLeft,
             /* Current page. */
             void *pvSeg;
             paSegs[iSeg].pvSeg = pvSeg = pPages->papvPages[iPage];
-            size_t cbSeg = PAGE_SIZE - ((uintptr_t)pvSeg & PAGE_OFFSET_MASK);
+            uint32_t cbSeg = PAGE_SIZE - (uint32_t)((uintptr_t)pvSeg & PAGE_OFFSET_MASK);
             iPage++;
 
             /* Adjacent to the next page? */
